Initial version of the contextual settings dropdown
