+++ packages/playground/website/src/components/browser-chrome/index.tsx
@@ -9,6 +9,10 @@ import {
 	useActiveSite,
 } from '../../lib/state/redux/store';
 import { SyncLocalFilesButton } from '../sync-local-files-button';
+import { Dropdown, Icon } from '@wordpress/components';
+import { cog } from '@wordpress/icons';
+import Button from '../button';
+import { ActiveSiteSettingsForm } from '../site-manager/site-settings-form';
 
 interface BrowserChromeProps {
 	children?: React.ReactNode;
@@ -57,6 +61,41 @@ export default function BrowserChrome({
 					</div>
 
 					<div className={css.toolbarButtons}>
+						<Dropdown
+							className="my-container-class-name"
+							contentClassName="my-dropdown-content-classname"
+							popoverProps={{ placement: 'bottom-start' }}
+							renderToggle={({ isOpen, onToggle }) => (
+								<Button
+									variant="browser-chrome"
+									onClick={onToggle}
+									aria-expanded={isOpen}
+									style={{
+										padding: '0 10px',
+										fill: '#FFF',
+										alignItems: 'center',
+										display: 'flex',
+									}}
+								>
+									<Icon icon={cog} />
+								</Button>
+							)}
+							renderContent={({ onClose }) => (
+								<div style={{ width: 400, padding: 0 }}>
+									<div
+										className={css.padded}
+										style={{ paddingBottom: 0 }}
+									>
+										<h2 style={{ margin: 0 }}>
+											Playground settings
+										</h2>
+									</div>
+									<ActiveSiteSettingsForm
+										onSubmit={onClose}
+									/>
+								</div>
+							)}
+						/>
 						{activeSite?.metadata?.storage === 'local-fs' ? (
 							<SyncLocalFilesButton />
 						) : null}
