+++ packages/playground/website/src/components/site-manager/icons.tsx
@@ -93,13 +93,14 @@ export const FolderIcon = (
 	</svg>
 );
 
-export const ClockIcon = (
+export const ClockIcon = (props?: React.SVGProps<SVGSVGElement>) => (
 	<svg
 		xmlns="http://www.w3.org/2000/svg"
 		width="14"
 		height="14"
 		viewBox="0 0 14 14"
 		fill="none"
+		{...props}
 	>
 		<path
 			fillRule="evenodd"
