+++ packages/playground/website/src/components/site-manager/sidebar/style.module.css
@@ -32,7 +32,7 @@
 .sidebar-label {
 	color: #949494 !important;
 	text-transform: uppercase;
-	padding: 0 12px 8px;
+	padding: 8px 12px 8px;
 	font-size: 12px !important;
 }
