+++ packages/playground/website/src/components/site-manager/site-settings-form/style.module.css
@@ -1,5 +1,13 @@
+.site-title {
+	margin: 0;
+}
+
 .grow-children > * {
 	flex-grow: 1;
+}
+
+:global(.components-popover__content:has(.site-settings-form)) {
+	padding: 0;
 }
 
 .invalid-input :global(.components-input-control__backdrop) {
@@ -8,4 +16,39 @@
 
 .invalid-input :global(.components-base-control__help) {
 	color: #e26f56;
+}
+
+.form-section {
+	padding: 24px;
+}
+
+.form-section p {
+	margin: 0;
+}
+
+.form-section :global(.components-select-control > :first-child) {
+	width: 50%;
+}
+.form-section :global(.components-select-control > :first-child label) {
+	text-transform: none;
+	font-weight: normal;
+	font-size: 14px;
+	color: #757575;
+}
+.form-section :global(.components-select-control > :nth-child(2)) {
+	width: auto;
+	flex-shrink: 1;
+}
+
+.notice {
+	background-color: #f3f6fd;
+}
+
+.footer {
+	background-color: #f3f6fd;
+	button {
+		width: 100%;
+		align-self: center;
+		justify-content: center;
+	}
 }
