+++ packages/playground/website/src/components/site-manager/storage-type/index.tsx
@@ -1,5 +1,5 @@
 import { Icon } from '@wordpress/components';
-import { ClockIcon, FolderIcon, LayoutIcon } from '../icons';
+import { clockIcon, FolderIcon, LayoutIcon } from '../icons';
 import css from './style.module.css';
 import { SiteStorageType } from '../../../lib/site-metadata';
 
@@ -22,7 +22,7 @@ export function StorageType({ type }: { type: SiteStorageType }) {
 		case 'none':
 			return (
 				<div className={css.storageType}>
-					<Icon size={16} icon={ClockIcon} />
+					<Icon size={16} icon={clockIcon} />
 					<span>Temporary</span>
 				</div>
 			);
